+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.12.0-rc.21",
-  genVersion: "2.917.0",
-  userAgent: "speakeasy-sdk/typescript 4.12.0-rc.21 2.917.0 2.0 @orq-ai/node",
+  sdkVersion: "4.12.0-rc.22",
+  genVersion: "2.918.1",
+  userAgent: "speakeasy-sdk/typescript 4.12.0-rc.22 2.918.1 2.0 @orq-ai/node",
 } as const;