+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4040,7 +4040,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KXG07W91H7K5A5ZC2ERG56M6"),
+  id: z.string().default("01KXGBS95RSZ68D7B5CP3A7731"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),