+++ packages/orq-rc/RELEASES.md
@@ -10769,3 +10769,13 @@ Based on:
 - [typescript v4.12.0-rc.21] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.21 - packages/orq-rc
+
+## 2026-07-14 13:09:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.22] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.22 - packages/orq-rc